+++ NEWS.md
@@ -19,6 +19,7 @@ Bug fixes:
 + Properly handle `CompileError` (that are not the subclass `ParseError`). CompileError was added in PHP 7.3.
   (Phan now logs these the same way it would log other syntax errors, instead of treating this like an unexpected Error.)
 + Make sure that private methods that are generators, that are inherited from a trait, aren't treated like a `void`.
++ Fix a crash analyzing a dynamic call to a static method, which occurred when dead code detection or reference tracking was enabled. (#1889)
 
 12 Aug 2018, Phan 1.0.0
 -----------------------